block settings update
